backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
google-cloud-bigquery>=3.9.0
python-dotenv>=1.0.0